+++ packages/NewFlashCardsEdit/edit.tsx
@@ -1,16 +1,16 @@
-import style from "./style.module.css";
+import style from './style.module.css';
 // import style from './style.module.css'
 // import './style.module.css'
 
-import { FileUploadButton } from "@/components/index";
+import { FileUploadButton } from '@/components/index';
 import {
   ModuleCourseContentComponentContentType,
   LocaleTypes,
   ModuleCodingContentType,
   InsertMultipleContentVals,
   GenerateNewTextKeysArgs,
-} from "@/funcs/courseModuleTypes";
-import { useState } from "react";
+} from '@/funcs/courseModuleTypes';
+import { useState } from 'react';
 
 interface FlashCardsStructureProps {
   textId?: string;
@@ -50,7 +50,7 @@ interface EditProps {
   // crudCodeSubModuleByIndex: (sectionIndex: number, componentIndex: number, action: string, data: any) => void;
   generateNewTextKeys: (args: GenerateNewTextKeysArgs) => string[];
   insertMultipleContentVals?: (args: InsertMultipleContentVals) => void;
-  deleteMultipleContentVals?: (list: InsertMultipleContentVals["list"]) => void;
+  deleteMultipleContentVals?: (list: InsertMultipleContentVals['list']) => void;
 }
 
 const Edit = (props: EditProps) => {
@@ -65,7 +65,7 @@ const Edit = (props: EditProps) => {
     deleteMultipleContentVals,
   } = props;
   const { textId, props: compProps } = structureComponent;
-  const [lang, country] = props.locale.split("-");
+  const [lang, country] = props.locale.split('-');
   // const identifier = `${props.sectionIndex}_${props.componentIndex}`;
 
   type FlashCardsFlipStates = { [key: string]: boolean };
@@ -73,9 +73,9 @@ const Edit = (props: EditProps) => {
     useState<FlashCardsFlipStates>({});
 
   const [activeFlashcardUploader, setActiveFlashcardUploader] =
-    useState<string>("");
+    useState<string>('');
 
-  const { flashcards } = compProps as FlashCardsStructureProps["props"];
+  const { flashcards } = compProps as FlashCardsStructureProps['props'];
 
   /** If code update needed: */
   /**
@@ -97,9 +97,9 @@ const Edit = (props: EditProps) => {
     newFlashcards[index].front_image = imageUrl;
 
     setStructureVal({
-      key: "flashcards",
+      key: 'flashcards',
       value: newFlashcards,
-      mainKey: "props",
+      mainKey: 'props',
       sectionIndex: props.sectionIndex,
       currentComponentIndex: props.componentIndex,
     });
@@ -108,7 +108,7 @@ const Edit = (props: EditProps) => {
   const removeFlashCard = (index: number) => {
     if (!flashcards || flashcards.length === 0) return;
     let contentKeysToRemove = [] as string[];
-    flashcards.forEach((flashcard) => {
+    flashcards.forEach(flashcard => {
       contentKeysToRemove.push(flashcard.front);
       contentKeysToRemove.push(flashcard.back);
     });
@@ -116,40 +116,40 @@ const Edit = (props: EditProps) => {
     newFlashcards.splice(index, 1);
     // fix flascardstates
     setStructureVal({
-      key: "flashcards",
+      key: 'flashcards',
       value: newFlashcards,
-      mainKey: "props",
+      mainKey: 'props',
       sectionIndex: props.sectionIndex,
       currentComponentIndex: props.componentIndex,
     });
     // remove keys from setContentVal
     if (!deleteMultipleContentVals) return;
     deleteMultipleContentVals(
-      contentKeysToRemove.map((textId) => ({
+      contentKeysToRemove.map(textId => ({
         textId,
-        value: "",
-      }))
+        value: '',
+      })),
     );
   };
   const createNewFlashCard = () => {
     let newContentKeys = generateNewTextKeys({
       numberOfKeys: 2,
       prefixObject: {
-        0: "flashcard_front",
-        1: "flashcard_back",
+        0: 'flashcard_front',
+        1: 'flashcard_back',
       },
     });
     const newFlashcard = {
       id: newContentKeys[0],
       front: newContentKeys[0],
-      front_image: "",
+      front_image: '',
       back: newContentKeys[1],
     };
     const newFlashcards = [...flashcards, newFlashcard];
     setStructureVal({
-      key: "flashcards",
+      key: 'flashcards',
       value: newFlashcards,
-      mainKey: "props",
+      mainKey: 'props',
       sectionIndex: props.sectionIndex,
       currentComponentIndex: props.componentIndex,
     });
@@ -160,11 +160,11 @@ const Edit = (props: EditProps) => {
       list: [
         {
           textId: newContentKeys[0],
-          value: "",
+          value: '',
         },
         {
           textId: newContentKeys[1],
-          value: "",
+          value: '',
         },
       ],
       lang,
@@ -191,88 +191,85 @@ const Edit = (props: EditProps) => {
 
   return (
     <div
-      className="p-4 rounded-md block"
+      className='p-4 rounded-md block'
       style={{
-        padding: "20px",
-        direction: "ltr",
-      }}
-    >
-      <span className="text-gray-500 italic ">Flashcards</span>
-      <div className="relative z-[1] bg-white">
-        <div className={`grid grid-cols-1 gap-4 ${style["flashcard_wrapper"]}`}>
+        padding: '20px',
+        direction: 'ltr',
+      }}>
+      <span className='text-gray-500 italic '>Flashcards</span>
+      <div className='relative z-[1] bg-white'>
+        <div className={`grid grid-cols-1 gap-4 ${style['flashcard_wrapper']}`}>
           {flashcards.map((flashcard, index) => {
             return (
               <div
                 key={flashcard.id}
-                className={`col-span-1 ${style["flip"]} mb-5`}
-              >
-                <div className="absolute z-30 active:h-full m-auto w-full hover:bg-amber-500/20 rounded-md active:scale-95 transition-all">
+                className={`col-span-1 ${style['flip']} mb-5`}>
+                <div className='absolute z-30 active:h-full m-auto w-full hover:bg-amber-500/20 rounded-md active:scale-95 transition-all'>
                   <button
-                    onClick={(e) => {
+                    onClick={e => {
                       e.stopPropagation();
                       removeFlashCard(index);
                     }}
-                    className="w-2xs m-auto block bg-gray-100 cursor-pointer hover:text-amber-600 border  rounded-md"
-                  >
+                    className='w-2xs m-auto block bg-gray-100 cursor-pointer hover:text-amber-600 border  rounded-md'>
                     Remove
                   </button>
                 </div>
-                <div className={style["card"]}>
+                <div className={style['card']}>
                   {/* front */}
                   <div
                     className={
                       flashcardsStates[flashcard.id]
-                        ? `${style["flipped"]} ${style["front"]}`
-                        : style["front"]
+                        ? `${style['flipped']} ${style['front']}`
+                        : style['front']
                     }
-                    role="button"
+                    role='button'
                     tabIndex={0}
-                    onClick={() => flipCard(flashcard.id)}
-                  >
-                    <div className="flex flex-col items-center justify-center bg-gradient-to-bl from-blue-50 to-green-50">
-                      <span className="relative mb-3">
+                    onClick={() => flipCard(flashcard.id)}>
+                    <div className='flex flex-col items-center justify-center bg-gradient-to-bl from-blue-50 to-green-50'>
+                      <span className='relative mb-3'>
                         {flashcard.front_image ? (
-                          <img src={flashcard.front_image} alt="front" />
+                          <img src={flashcard.front_image} alt='front' />
                         ) : (
-                          <span className="text-gray-500 italic">No image</span>
+                          <span className='text-gray-500 italic'>No image</span>
                         )}
-                        <span className="flex flex-row ">
+                        <span className='flex flex-row '>
                           {flashcard.front_image && (
                             <button
-                              onClick={(e) => {
+                              onClick={e => {
                                 e.stopPropagation();
-                                setFlashCardImage(index, "");
+                                setFlashCardImage(index, '');
                               }}
-                              className="p-1 border rounded-md mb-2 cursor-pointer hover:text-gray-600"
-                            >
+                              className='p-1 border rounded-md mb-2 cursor-pointer hover:text-gray-600'>
                               remove image
                             </button>
                           )}
                           <span>
                             <button
-                              onClick={(e) => {
+                              onClick={e => {
                                 e.stopPropagation();
                                 setActiveFlashcardUploader(flashcard.id);
                               }}
-                              className="ml-2 p-1 border rounded-md mb-2 cursor-pointer hover:text-gray-600"
-                            >
+                              className='ml-2 p-1 border rounded-md mb-2 cursor-pointer hover:text-gray-600'>
                               {flashcard.front_image
-                                ? "Change image"
-                                : "Add image"}
+                                ? 'Change image'
+                                : 'Add image'}
                             </button>
 
                             {activeFlashcardUploader === flashcard.id ? (
                               <FileUploadButton
+                                isOpen={(o: any) => {
+                                  if (!o) setActiveFlashcardUploader('');
+                                }}
                                 modalOnly={true}
                                 text={
                                   flashcard.front_image
-                                    ? "Change image"
-                                    : "Add image"
+                                    ? 'Change image'
+                                    : 'Add image'
                                 }
                                 options={uploadOptions}
                                 onComplete={(data: any) => {
                                   setFlashCardImage(index, data.url);
-                                  setActiveFlashcardUploader("");
+                                  setActiveFlashcardUploader('');
                                 }}
                               />
                             ) : null}
@@ -281,22 +278,21 @@ const Edit = (props: EditProps) => {
                       </span>
 
                       <textarea
-                        placeholder="Write front here"
-                        onClick={(e) => {
+                        placeholder='Write front here'
+                        onClick={e => {
                           e.stopPropagation();
                         }}
-                        onChange={(e) => {
+                        onChange={e => {
                           if (!props.setContentVal) return;
                           props.setContentVal({
                             textId: flashcard.front,
                             value: e.target.value,
-                            lang: props.locale.split("-")[0],
-                            country: props.locale.split("-")[1],
+                            lang: props.locale.split('-')[0],
+                            country: props.locale.split('-')[1],
                           });
                         }}
                         value={textData.getText(flashcard.front)}
-                        className="w-full mx-2 border text-center"
-                      ></textarea>
+                        className='w-full mx-2 border text-center'></textarea>
                     </div>
                   </div>
 
@@ -304,34 +300,32 @@ const Edit = (props: EditProps) => {
                   <div
                     className={
                       flashcardsStates[flashcard.id]
-                        ? `${style["flipped"]} ${style["back"]}`
-                        : style["back"]
+                        ? `${style['flipped']} ${style['back']}`
+                        : style['back']
                     }
-                    role="button"
+                    role='button'
                     tabIndex={0}
-                    onClick={() => flipCard(flashcard.id)}
-                  >
-                    <div className="flex flex-col items-center justify-center bg-gradient-to-bl from-yellow-50 to-blue-50">
+                    onClick={() => flipCard(flashcard.id)}>
+                    <div className='flex flex-col items-center justify-center bg-gradient-to-bl from-yellow-50 to-blue-50'>
                       {flashcard.front && (
                         <h2>{textData.getText(flashcard.front)}</h2>
                       )}
                       <textarea
-                        onChange={(e) => {
+                        onChange={e => {
                           if (!props.setContentVal) return;
                           props.setContentVal({
                             textId: flashcard.back,
                             value: e.target.value,
-                            lang: props.locale.split("-")[0],
-                            country: props.locale.split("-")[1],
+                            lang: props.locale.split('-')[0],
+                            country: props.locale.split('-')[1],
                           });
                         }}
-                        placeholder="Write back here"
-                        onClick={(e) => {
+                        placeholder='Write back here'
+                        onClick={e => {
                           e.stopPropagation();
                         }}
                         value={textData.getText(flashcard.back)}
-                        className="w-full mx-2 border text-center h-45"
-                      ></textarea>
+                        className='w-full mx-2 border text-center h-45'></textarea>
                     </div>
                   </div>
                 </div>
@@ -342,8 +336,7 @@ const Edit = (props: EditProps) => {
       </div>
       <button
         onClick={() => createNewFlashCard()}
-        className="w-full m-auto block bg-gray-100 cursor-pointer hover:text-amber-600 border  rounded-md"
-      >
+        className='w-full m-auto block bg-gray-100 cursor-pointer hover:text-amber-600 border  rounded-md'>
         Add Flashcard
       </button>
     </div>