+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Release 2 - March 13, 2025
+* New matching pairs components added
+* New Flash Cards component added
+
 # Release 1 - March 6, 2025
 * Added Example React iClasser Package
 * Wrote Readme and more docs